requirements.txt
python-dotenv==16.4.5
pandas==2.2.2
SQLAlchemy==2.0.30